+++ packages/pangea-sdk/src/main/java/cloud/pangeacyber/pangea/authn/models/AgreementListOrderBy.java
@@ -1,0 +1,29 @@
+package cloud.pangeacyber.pangea.authn.models;
+
+import com.fasterxml.jackson.annotation.JsonValue;
+
+public enum AgreementListOrderBy {
+	ID("id"),
+	CREATED_AT("created_at"),
+	NAME("name"),
+	TEXT("text");
+
+	private final String text;
+
+	AgreementListOrderBy(String text) {
+		this.text = text;
+	}
+
+	/* (non-Javadoc)
+	 * @see java.lang.Enum#toString()
+	 */
+	@Override
+	public String toString() {
+		return text;
+	}
+
+	@JsonValue
+	final String value() {
+		return text;
+	}
+}