+++ packages/pangea-sdk/src/main/java/cloud/pangeacyber/pangea/authn/models/Filter.java
@@ -1,7 +1,0 @@
-package cloud.pangeacyber.pangea.authn.models;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import java.util.HashMap;
-
-@JsonIgnoreProperties(ignoreUnknown = true)
-public class Filter extends HashMap<String, Object> {}