+++ packages/pangea-sdk/src/main/java/cloud/pangeacyber/pangea/file_scan/responses/FileScanResponse.java
@@ -1,0 +1,6 @@
+package cloud.pangeacyber.pangea.file_scan.responses;
+
+import cloud.pangeacyber.pangea.Response;
+import cloud.pangeacyber.pangea.file_scan.results.FileScanResult;
+
+public final class FileScanResponse extends Response<FileScanResult> {}