+++ packages/pangea-sdk/src/main/java/cloud/pangeacyber/pangea/intel/IPIntelClient.java
@@ -10,10 +10,9 @@ import cloud.pangeacyber.pangea.intel.responses.*;
 public class IPIntelClient extends BaseClient {
 
 	public static String serviceName = "ip-intel";
-	private static final boolean supportMultiConfig = false;
 
 	public IPIntelClient(Builder builder) {
-		super(builder, serviceName, supportMultiConfig);
+		super(builder, serviceName);
 	}
 
 	public static class Builder extends BaseClient.Builder<Builder> {
@@ -28,61 +27,120 @@ public class IPIntelClient extends BaseClient {
 	}
 
 	/**
-	 * Geolocate - provider, verbose, raw
-	 * @pangea.description Retrieve information about the location of an IP address
+	 * Geolocate
+	 * @pangea.description Retrieve location information associated with an IP address.
+	 * @pangea.operationId ip_intel_post_v1_geolocate
 	 * @param request
 	 * @return IPGeolocateResponse
 	 * @throws PangeaException
 	 * @throws PangeaAPIException
+	 * @pangea.code
+	 * {@code
+	 * IPGeolocateRequest request = new IPGeolocateRequest
+	 * 	.Builder("93.231.182.110")
+	 * 	.provider("digitalelement")
+	 * 	.verbose(true)
+	 * 	.raw(true)
+	 * 	.build();
+	 *
+	 * IPGeolocateResponse response = client.geolocate(request);
+	 * }
 	 */
 	public IPGeolocateResponse geolocate(IPGeolocateRequest request) throws PangeaException, PangeaAPIException {
 		return post("/v1/geolocate", request, IPGeolocateResponse.class);
 	}
 
 	/**
-	 * Domain - ip, provider, verbose, raw
-	 * @pangea.description Retrieve the domain name associated with an IP address
+	 * Domain
+	 * @pangea.description Retrieve the domain name associated with an IP address.
+	 * @pangea.operationId ip_intel_post_v1_domain
 	 * @param request
 	 * @return IpDomainResponse
 	 * @throws PangeaException
 	 * @throws PangeaAPIException
+	 * @pangea.code
+	 * {@code
+	 * IPDomainRequest request = new IPDomainRequest
+	 * 	.Builder("93.231.182.110")
+	 * 	.provider("digitalelement")
+	 * 	.verbose(true)
+	 * 	.raw(true)
+	 * 	.build();
+	 *
+	 * IPDomainResponse response = client.getDomain(request);
+	 * }
 	 */
 	public IPDomainResponse getDomain(IPDomainRequest request) throws PangeaException, PangeaAPIException {
 		return post("/v1/domain", request, IPDomainResponse.class);
 	}
 
 	/**
-	 * VPN - ip, provider, verbose, raw
-	 * @pangea.description Determine if an IP address is provided by a VPN service
+	 * VPN
+	 * @pangea.description Determine if an IP address originates from a VPN.
+	 * @pangea.operationId ip_intel_post_v1_vpn
 	 * @param request
 	 * @return IpVPNResponse
 	 * @throws PangeaException
 	 * @throws PangeaAPIException
+	 * @pangea.code
+	 * {@code
+	 * IPVPNRequest request = new IPVPNRequest
+	 * 	.Builder("93.231.182.110")
+	 * 	.provider("digitalelement")
+	 * 	.verbose(true)
+	 * 	.raw(true)
+	 * 	.build();
+	 *
+	 * IPVPNResponse response = client.isVPN(request);
+	 * }
 	 */
 	public IPVPNResponse isVPN(IPVPNRequest request) throws PangeaException, PangeaAPIException {
 		return post("/v1/vpn", request, IPVPNResponse.class);
 	}
 
 	/**
-	 * Proxy - provider, verbose, raw
-	 * @pangea.description Determine if an IP address is provided by a proxy service
+	 * Proxy
+	 * @pangea.description Determine if an IP address originates from a proxy.
+	 * @pangea.operationId ip_intel_post_v1_proxy
 	 * @param request
 	 * @return IpProxyResponse
 	 * @throws PangeaException
 	 * @throws PangeaAPIException
+	 * @pangea.code
+	 * {@code
+	 * IPProxyRequest request = new IPProxyRequest
+	 * 	.Builder("34.201.32.172")
+	 * 	.provider("digitalelement")
+	 * 	.verbose(true)
+	 * 	.raw(true)
+	 * 	.build();
+	 *
+	 * IPProxyResponse response = client.isProxy(request);
+	 * }
 	 */
 	public IPProxyResponse isProxy(IPProxyRequest request) throws PangeaException, PangeaAPIException {
 		return post("/v1/proxy", request, IPProxyResponse.class);
 	}
 
 	/**
-	 * Reputation - provider, verbose, raw
+	 * Reputation
 	 * @pangea.description Retrieve a reputation score for an IP address from a provider, including an optional detailed report.
+	 * @pangea.operationId ip_intel_post_v1_reputation
 	 * @param request
 	 * @return IpReputationResponse
 	 * @throws PangeaException
 	 * @throws PangeaAPIException
 	 * @pangea.code
+	 * {@code
+	 * IPReputationRequest request = new IPReputationRequest
+	 * 	.Builder("93.231.182.110")
+	 * 	.provider("crowdstrike")
+	 * 	.verbose(true)
+	 * 	.raw(true)
+	 * 	.build();
+	 *
+	 * IPReputationResponse response = client.reputation(request);
+	 * }
 	 */
 	public IPReputationResponse reputation(IPReputationRequest request) throws PangeaException, PangeaAPIException {
 		return post("/v1/reputation", request, IPReputationResponse.class);