+++ packages/pangea-sdk/src/test/java/cloud/pangeacyber/pangea/authn/ITAuthNTest.java
@@ -11,6 +11,8 @@ import cloud.pangeacyber.pangea.authn.models.*;
 import cloud.pangeacyber.pangea.authn.requests.*;
 import cloud.pangeacyber.pangea.authn.responses.*;
 import cloud.pangeacyber.pangea.exceptions.*;
+import java.time.LocalDateTime;
+import java.time.format.DateTimeFormatter;
 import java.util.HashMap;
 import java.util.Random;
 import org.junit.Before;
@@ -35,11 +37,14 @@ public class ITAuthNTest {
 	private static final Profile profileOld = new Profile();
 	private static final Profile profileNew = new Profile();
 	private static String userID = ""; // Will be set once user is created
+	String time;
 
 	@Before
 	public void setUp() throws ConfigException {
 		this.cfg = Config.fromIntegrationEnvironment(environment);
 		client = new AuthNClient.Builder(cfg).build();
+		DateTimeFormatter dtf = DateTimeFormatter.ofPattern("yyyyMMdd_HHmmss");
+		time = dtf.format(LocalDateTime.now());
 
 		profileOld.put("name", "User name");
 		profileOld.put("country", "Argentina");
@@ -207,11 +212,13 @@ public class ITAuthNTest {
 			assertNotNull(loginResp.getResult().getRefreshToken());
 			String token = loginResp.getResult().getActiveToken().getToken();
 
+			FilterSessionList filter = new FilterSessionList();
+
 			// List client sessions
 			ClientSessionListResponse listResp = client
 				.client()
 				.session()
-				.list(new ClientSessionListRequest.Builder(token).build());
+				.list(new ClientSessionListRequest.Builder(token).setFilter(filter).build());
 			assertTrue(listResp.isOk());
 			assertTrue(listResp.getResult().getSessions().length > 0);
 
@@ -246,8 +253,12 @@ public class ITAuthNTest {
 			assertNotNull(loginResp.getResult().getRefreshToken());
 			String token = loginResp.getResult().getActiveToken().getToken();
 
+			FilterSessionList filter = new FilterSessionList();
+
 			// Session list
-			SessionListResponse listResp = client.session().list(new SessionListRequest.Builder().build());
+			SessionListResponse listResp = client
+				.session()
+				.list(new SessionListRequest.Builder().setFilter(filter).build());
 			assertTrue(listResp.isOk());
 			assertTrue(listResp.getResult().getSessions().length > 0);
 
@@ -322,12 +333,14 @@ public class ITAuthNTest {
 			assertTrue(false);
 		}
 
+		FilterUserInviteList filter = new FilterUserInviteList();
+
 		try {
 			// List users invites
 			UserInviteListResponse inviteListResp1 = client
 				.user()
 				.invite()
-				.list(new UserInviteListRequest.Builder().build());
+				.list(new UserInviteListRequest.Builder().setFilter(filter).build());
 			assertTrue(inviteListResp1.isOk());
 			assertNotNull(inviteListResp1.getResult().getInvites());
 			assertTrue(inviteListResp1.getResult().getInvites().length > 0);
@@ -345,8 +358,12 @@ public class ITAuthNTest {
 
 	@Test
 	public void testE_ListUsers() throws PangeaException, PangeaAPIException {
+		FilterUserList filter = new FilterUserList();
+
 		try {
-			UserListResponse userListResp1 = client.user().list(new UserListRequest.Builder().build());
+			UserListResponse userListResp1 = client
+				.user()
+				.list(new UserListRequest.Builder().setFilter(filter).build());
 			assertTrue(userListResp1.isOk());
 			assertTrue(userListResp1.getResult().getUsers().length > 0);
 
@@ -359,6 +376,80 @@ public class ITAuthNTest {
 					System.out.println(e.toString());
 				}
 			}
+		} catch (PangeaAPIException e) {
+			System.out.println(e.toString());
+			assertTrue(false);
+		}
+	}
+
+	public void agreementsCycle(AgreementType type) throws PangeaAPIException, PangeaException {
+		String name = type + "_" + System.currentTimeMillis();
+		String text = "This is agreement text";
+		boolean active = false;
+
+		// Create agreement
+		AgreementCreateResponse createResponse = client
+			.agreements()
+			.create(new AgreementCreateRequest.Builder(type, name, text).setActive(active).build());
+		assertEquals(type.toString(), createResponse.getResult().getType());
+		assertEquals(name, createResponse.getResult().getName());
+		assertEquals(text, createResponse.getResult().getText());
+		assertEquals(active, createResponse.getResult().isActive());
+		String id = createResponse.getResult().getID();
+		assertNotNull(id);
+
+		// Update agreement
+		String new_name = name + "_v2";
+		String new_text = text + " v2";
+
+		AgreementUpdateResponse updateResponse = client
+			.agreements()
+			.update(
+				new AgreementUpdateRequest.Builder(type, id)
+					.setName(new_name)
+					.setText(new_text)
+					.setActive(active)
+					.build()
+			);
+		assertEquals(new_name, updateResponse.getResult().getName());
+		assertEquals(new_text, updateResponse.getResult().getText());
+		assertEquals(active, updateResponse.getResult().isActive());
+
+		// List
+		FilterAgreementList filter = new FilterAgreementList();
+		AgreementListResponse listResponse = client
+			.agreements()
+			.list(new AgreementListRequest.Builder().setFilter(filter).build());
+		assertTrue(listResponse.getResult().getCount() > 0);
+		assertTrue(listResponse.getResult().getAgreements().length > 0);
+		int count = listResponse.getResult().getCount();
+
+		// Delete
+		AgreementDeleteResponse deleteResponse = client
+			.agreements()
+			.delete(new AgreementDeleteRequest.Builder(type, id).build());
+
+		// List again
+		AgreementListResponse listResponseAfterDelete = client
+			.agreements()
+			.list(new AgreementListRequest.Builder().build());
+		assertEquals(count - 1, (int) listResponseAfterDelete.getResult().getCount());
+	}
+
+	@Test
+	public void test_AgreementsCycleEULA() throws PangeaException, PangeaAPIException {
+		try {
+			agreementsCycle(AgreementType.EULA);
+		} catch (PangeaAPIException e) {
+			System.out.println(e.toString());
+			assertTrue(false);
+		}
+	}
+
+	@Test
+	public void test_AgreementsCyclePrivacyPolicy() throws PangeaException, PangeaAPIException {
+		try {
+			agreementsCycle(AgreementType.PRIVACY_POLICY);
 		} catch (PangeaAPIException e) {
 			System.out.println(e.toString());
 			assertTrue(false);