+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.2.0] - 2023-09-05
+
+# Added
+
+- Redact rulesets field support 
+- FileScan service support
+
+
 ## [2.1.0] - 2023-07-14
 
 # Added
@@ -182,7 +190,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Redact client
 
 
-[unreleased]: https://github.com/pangeacyber/pangea-java/compare/v2.1.0...main
+[unreleased]: https://github.com/pangeacyber/pangea-java/compare/v2.2.0...main
+[2.2.0]: https://github.com/pangeacyber/pangea-java/compare/v2.1.0...v2.2.0
 [2.1.0]: https://github.com/pangeacyber/pangea-java/compare/v2.0.0...v2.1.0
 [2.0.0]: https://github.com/pangeacyber/pangea-java/compare/v1.10.0...v2.0.0
 [1.10.0]: https://github.com/pangeacyber/pangea-java/compare/v1.9.1...v1.10.0